feat: add typescript for Gatsby and update eslint rules
