+++ package.json
@@ -36,6 +36,7 @@
         "@commitlint/config-conventional": "^8.0.0",
         "@storybook/preset-typescript": "^1.1.0",
         "@storybook/react": "^5.1.9",
+        "@types/react": "^16.8.23",
         "@typescript-eslint/eslint-plugin": "^1.11.0",
         "@typescript-eslint/parser": "^1.11.0",
         "babel-loader": "^8.0.6",
@@ -44,6 +45,7 @@
         "cz-lerna-changelog": "^2.0.2",
         "eslint": "^6.0.1",
         "eslint-config-prettier": "^6.0.0",
+        "eslint-config-react": "^1.1.7",
         "eslint-import-resolver-lerna": "^1.1.0",
         "eslint-plugin-import": "^2.18.0",
         "eslint-plugin-jest": "^22.7.2",
