+++ packages/website/package.json
@@ -1,46 +1,51 @@
 {
-  "name": "platoon-website",
-  "private": true,
-  "description": "A simple starter to get up and developing quickly with Gatsby",
-  "version": "0.1.0",
-  "author": "Benoit Deziel <ben@benoitdeziel.com>",
-  "dependencies": {
-    "@mdx-js/mdx": "^1.0.21",
-    "@mdx-js/react": "^1.0.21",
-    "gatsby": "^2.13.6",
-    "gatsby-image": "^2.2.4",
-    "gatsby-plugin-manifest": "^2.2.1",
-    "gatsby-plugin-mdx": "^1.0.7",
-    "gatsby-plugin-offline": "^2.2.1",
-    "gatsby-plugin-react-helmet": "^3.1.0",
-    "gatsby-plugin-sharp": "^2.2.3",
-    "gatsby-source-filesystem": "^2.1.2",
-    "gatsby-transformer-sharp": "^2.2.1",
-    "prop-types": "^15.7.2",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "react-helmet": "^5.2.1"
-  },
-  "devDependencies": {
-    "prettier": "^1.18.2"
-  },
-  "keywords": [
-    "gatsby"
-  ],
-  "license": "MIT",
-  "scripts": {
-    "build": "gatsby build",
-    "develop": "gatsby develop",
-    "format": "prettier --write src/**/*.{js,jsx}",
-    "start": "npm run develop",
-    "serve": "gatsby serve",
-    "test": "echo \"Write tests! -> https://gatsby.dev/unit-testing\""
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/gatsbyjs/gatsby-starter-default"
-  },
-  "bugs": {
-    "url": "https://github.com/gatsbyjs/gatsby/issues"
-  }
+    "name": "platoon-website",
+    "private": true,
+    "description": "A simple starter to get up and developing quickly with Gatsby",
+    "version": "0.1.0",
+    "author": "Benoit Deziel <ben@benoitdeziel.com>",
+    "dependencies": {
+        "@mdx-js/mdx": "^1.0.21",
+        "@mdx-js/react": "^1.0.21",
+        "gatsby": "^2.13.6",
+        "gatsby-image": "^2.2.4",
+        "gatsby-plugin-manifest": "^2.2.1",
+        "gatsby-plugin-mdx": "^1.0.7",
+        "gatsby-plugin-offline": "^2.2.1",
+        "gatsby-plugin-react-helmet": "^3.1.0",
+        "gatsby-plugin-sharp": "^2.2.3",
+        "gatsby-plugin-tslint": "^0.0.2",
+        "gatsby-plugin-typescript": "^2.1.0",
+        "gatsby-source-filesystem": "^2.1.2",
+        "gatsby-transformer-sharp": "^2.2.1",
+        "prop-types": "^15.7.2",
+        "react": "^16.8.6",
+        "react-dom": "^16.8.6",
+        "react-helmet": "^5.2.1"
+    },
+    "devDependencies": {
+        "@types/react": "^16.8.23",
+        "@types/react-dom": "^16.8.4",
+        "@types/react-helmet": "^5.0.8",
+        "typescript": "^3.5.2"
+    },
+    "keywords": [
+        "gatsby"
+    ],
+    "license": "MIT",
+    "scripts": {
+        "build": "gatsby build",
+        "develop": "gatsby develop",
+        "format": "prettier --write src/**/*.{js,jsx}",
+        "start": "npm run develop",
+        "serve": "gatsby serve",
+        "test": "echo \"Write tests! -> https://gatsby.dev/unit-testing\""
+    },
+    "repository": {
+        "type": "git",
+        "url": "https://github.com/gatsbyjs/gatsby-starter-default"
+    },
+    "bugs": {
+        "url": "https://github.com/gatsbyjs/gatsby/issues"
+    }
 }
