+++ packages/website/tsconfig.json
@@ -1,0 +1,24 @@
+{
+    "compilerOptions": {
+        "target": "es5",
+        "allowJs": true,
+        "lib": ["es2017", "dom"],
+        "skipLibCheck": true,
+        "esModuleInterop": true,
+        "allowSyntheticDefaultImports": true,
+        "strict": true,
+        "module": "esnext",
+        "moduleResolution": "node",
+        "isolatedModules": true,
+        "noEmit": true,
+        "jsx": "preserve",
+        "resolveJsonModule": true,
+        "rootDir": "src",
+        "baseUrl": ".",
+        "paths": {
+            "src/*": ["src/*"]
+        },
+        "forceConsistentCasingInFileNames": true
+    },
+    "include": ["src"]
+}
